# Align compaction output file boundaries to the next level ones