--- a/db/compaction/compaction.cc
+++ b/db/compaction/compaction.cc
@@ -270,9 +270,11 @@ Compaction::Compaction(
 
   // for the non-bottommost levels, it tries to build files match the target
   // file size, but not guaranteed. It could be 2x the size of the target size.
-  max_output_file_size_ = bottommost_level_ || grandparents_.empty()
-                              ? target_output_file_size_
-                              : 2 * target_output_file_size_;
+  max_output_file_size_ =
+      bottommost_level_ || grandparents_.empty() ||
+              !_immutable_options.level_compaction_dynamic_file_size
+          ? target_output_file_size_
+          : 2 * target_output_file_size_;
 
 #ifndef NDEBUG
   for (size_t i = 1; i < inputs_.size(); ++i) {