… HISTORY.md
… options/options_settable_test.cc
… options/options_test.cc