@mdcallag 2022-09-28
This change seems like it can have a big impact but the commit message content is too brief. Please include:
* a few more sentences about what was changed
* more detail on the perf test
* a description of whether or not this behavior depends on options. From the files I see there is a new option, that option should be explained in the commit message